ContextIQ Lite/requirements.txt
fastapi
uvicorn
openai
qdrant-client
langchain-text-splitters
playwright
beautifulsoup4
python-dotenv
qdrant-client
